requirements.txt


Gooey == 1.0.8.1
pymailtm == 1.0.2
random_password_generator == 2.2.0
random_username == 1.0.2
requests == 2.25.1
